--- adafruit_rsa/randnum.py
@@ -1,89 +1,0 @@
-# -*- coding: utf-8 -*-
-# SPDX-FileCopyrightText: 2011 Sybren A. Stüvel <sybren@stuvel.eu>
-#
-# SPDX-License-Identifier: Apache-2.0
-
-"""Functions for generating random numbers."""
-
-# Source inspired by code by Yesudeep Mangalapilly <yesudeep@gmail.com>
-
-import os
-
-from adafruit_rsa import common, transform
-from adafruit_rsa._compat import byte
-
-__version__ = "0.0.0-auto.0"
-__repo__ = "https://github.com/adafruit/Adafruit_CircuitPython_RSA.git"
-
-
-def read_random_bits(nbits):
-    """Reads 'nbits' random bits.
-
-    If nbits isn't a whole number of bytes, an extra byte will be appended with
-    only the lower bits set.
-    """
-
-    nbytes, rbits = divmod(nbits, 8)
-
-    # Get the random bytes
-    randomdata = os.urandom(nbytes)
-
-    # Add the remaining random bits
-    if rbits > 0:
-        randomvalue = ord(os.urandom(1))
-        randomvalue >>= 8 - rbits
-        randomdata = byte(randomvalue) + randomdata
-
-    return randomdata
-
-
-def read_random_int(nbits):
-    """Reads a random integer of approximately nbits bits."""
-
-    randomdata = read_random_bits(nbits)
-    value = transform.bytes2int(randomdata)
-
-    # Ensure that the number is large enough to just fill out the required
-    # number of bits.
-    value |= 1 << (nbits - 1)
-
-    return value
-
-
-def read_random_odd_int(nbits):
-    """Reads a random odd integer of approximately nbits bits.
-
-    >>> read_random_odd_int(512) & 1
-    1
-    """
-
-    value = read_random_int(nbits)
-
-    # Make sure it's odd
-    return value | 1
-
-
-def randint(maxvalue):
-    """Returns a random integer x with 1 <= x <= maxvalue
-
-    May take a very long time in specific situations. If maxvalue needs N bits
-    to store, the closer maxvalue is to (2 ** N) - 1, the faster this function
-    is.
-    """
-
-    bit_size = common.bit_size(maxvalue)
-
-    tries = 0
-    while True:
-        value = read_random_int(bit_size)
-        if value <= maxvalue:
-            break
-
-        if tries % 10 == 0 and tries:
-            # After a lot of tries to get the right number of bits but still
-            # smaller than maxvalue, decrease the number of bits by 1. That'll
-            # dramatically increase the chances to get a large enough number.
-            bit_size -= 1
-        tries += 1
-
-    return value
